Move ThreadFactoryNameDecorator -> ThrottledParallelCollector
